correção para incluir casas decimais no value

+++ backend/src/migrations/20241122040406-create-driver.js
@@ -32,7 +32,7 @@ module.exports = {
       },
       value: {
         allowNull: false,
-        type: Sequelize.DOUBLE
+        type: Sequelize.DOUBLE(10, 2)
       },
       kmMin: {
         allowNull: false,
